Replace old fries outputter.
