+++ src/main/scala/org/clulab/reach/extern/export/fries/FriesOutput.scala
@@ -12,6 +12,7 @@ import org.clulab.processors.Document
 import org.clulab.reach.context._
 import org.clulab.reach.display._
 import org.clulab.reach.extern.export._
+import org.clulab.reach.extern.export.MentionManager._
 import org.clulab.reach.grounding.KBResolution
 import org.clulab.reach.mentions._
 import JsonOutputter._
@@ -23,10 +24,11 @@ import scala.collection.mutable.ListBuffer
 
 /**
   * Defines classes and methods used to build and output the FRIES format.
-  *   Written by Mihai Surdeanu. 5/22/2015.
-  *   Last Modified: Move participant features to arguments, cleanup.
+  *   Written by: Mihai Surdeanu and Tom Hicks.
+  *   Last Modified: Revamp make argument method. More cleanups.
   */
 class FriesOutput extends JsonOutputter {
+
   // local type definitions:
   type IDed = scala.collection.mutable.HashMap[Mention, String]
   type CtxIDed = scala.collection.mutable.HashMap[ContextMap, String]
@@ -43,10 +45,8 @@ class FriesOutput extends JsonOutputter {
   // incrementing ID for numbering links
   protected val linkIdCntr = new IncrementingId()
 
-  var currEvent = ""
-
   // map incoming Reach context type names to FRIES-spec context type names
-  val contextNameMap = Map(
+  val contextNameMap = Map (
     "CellLine" -> "cell-line",
     "CellType" -> "cell-type",
     "TissueType" -> "tissue-type",
@@ -71,31 +71,14 @@ class FriesOutput extends JsonOutputter {
                        endTime:Date,
                        outFilePrefix:String): String = {
 
-    // Flatten mentions (per MITRE requirements), deduplicate, etc.
-    val sanitizedMentions = OutputDegrader.prepareForOutput(allMentions)
-
-    val otherMetaData = extractOtherMetaData(paperPassages)
-    val passageMap = passagesToMap(paperPassages) // map of FriesEntry, chunkId as key
-
-    val contextIdMap = new CtxIDed
-
-    val sentModel = sentencesToModel(paperId, sanitizedMentions, passageMap,
-                                     startTime, endTime, otherMetaData)
-
-    // entityMap: map from entity pointers to unique ids
-    val (entityModel, entityMap) = entitiesToModel(paperId, sanitizedMentions, passageMap, contextIdMap,
-                                                   startTime, endTime, otherMetaData)
-
-    // eventMap: map from entity pointers to unique ids
-    val (eventModel, eventMap) = eventsToModel(paperId, sanitizedMentions, passageMap, contextIdMap,
-                                               entityMap, startTime, endTime, otherMetaData)
+    val (sentModel, entityModel, eventModel, _) =
+      makeModels(paperId, allMentions, paperPassages, startTime, endTime)
 
     val uniModel:PropMap = new PropMap      // combine models into one
     uniModel("sentences") = sentModel
     uniModel("entities") = entityModel
     uniModel("events") = eventModel
-
-    writeJsonToString(uniModel)
+    writeJsonToString(uniModel)             // write out combined model
   }
 
 
@@ -111,26 +94,11 @@ class FriesOutput extends JsonOutputter {
                           endTime:Date,
                           outFilePrefix:String): Unit = {
 
-    // Flatten mentions (per MITRE requirements), deduplicate, etc.
-    val sanitizedMentions = OutputDegrader.prepareForOutput(allMentions)
+    val (sentModel, entityModel, eventModel, _) =
+      makeModels(paperId, allMentions, paperPassages, startTime, endTime)
 
-    val otherMetaData = extractOtherMetaData(paperPassages)
-    val passageMap = passagesToMap(paperPassages) // map of FriesEntry, chunkId as key
-
-    val contextIdMap = new CtxIDed
-
-    val sentModel = sentencesToModel(paperId, sanitizedMentions, passageMap,
-                                     startTime, endTime, otherMetaData)
     writeJsonToFile(sentModel, new File(outFilePrefix + ".uaz.sentences.json"))
-
-    // entityMap: map from entity pointers to unique ids
-    val (entityModel, entityMap) = entitiesToModel(paperId, sanitizedMentions, passageMap, contextIdMap,
-                                                   startTime, endTime, otherMetaData)
     writeJsonToFile(entityModel, new File(outFilePrefix + ".uaz.entities.json"))
-
-    // eventMap: map from entity pointers to unique ids
-    val (eventModel, eventMap) = eventsToModel(paperId, sanitizedMentions, passageMap, contextIdMap,
-                                               entityMap, startTime, endTime, otherMetaData)
     writeJsonToFile(eventModel, new File(outFilePrefix + ".uaz.events.json"))
   }
 
@@ -141,49 +109,80 @@ class FriesOutput extends JsonOutputter {
     * Each output file is prefixed with the given prefix string.
     */
   override def writeJSON (paperId:String,
-                 allMentions:Seq[Mention],
-                 paperPassages:Seq[FriesEntry],
-                 startTime:Date,
-                 endTime:Date,
-                 outFilePrefix:String,
-                 assemblyAPI: Assembler): Unit = {
+                          allMentions:Seq[Mention],
+                          paperPassages:Seq[FriesEntry],
+                          startTime:Date,
+                          endTime:Date,
+                          outFilePrefix:String,
+                          assemblyApi: Assembler): Unit = {
 
-    // Flatten mentions (per MITRE requirements), deduplicate, etc.
-    val sanitizedMentions = OutputDegrader.prepareForOutput(allMentions)
+    val (sentModel, entityModel, eventModel, assemblyModel) =
+      makeModels(paperId, allMentions, paperPassages, startTime, endTime, Some(assemblyApi))
 
-    val otherMetaData = extractOtherMetaData(paperPassages)
-    val passageMap = passagesToMap(paperPassages) // map of FriesEntry, chunkId as key
-
-    val contextIdMap = new CtxIDed
-
-    val sentModel = sentencesToModel(paperId, sanitizedMentions, passageMap,
-                                     startTime, endTime, otherMetaData)
     writeJsonToFile(sentModel, new File(outFilePrefix + ".uaz.sentences.json"))
-
-    // entityMap: map from entity pointers to unique ids
-    val (entityModel, entityMap) = entitiesToModel(paperId, sanitizedMentions, passageMap, contextIdMap,
-                                                   startTime, endTime, otherMetaData)
     writeJsonToFile(entityModel, new File(outFilePrefix + ".uaz.entities.json"))
-
-    // eventMap: map from entity pointers to unique ids
-    val (eventModel, eventMap) = eventsToModel(paperId, sanitizedMentions, passageMap,
-                                               contextIdMap, entityMap, startTime, endTime,
-                                               otherMetaData, Some(assemblyAPI))
     writeJsonToFile(eventModel, new File(outFilePrefix + ".uaz.events.json"))
-
-    val assemblyModel:PropMap = mkAssemblyModel(paperId, sanitizedMentions, passageMap, entityMap, eventMap,
-                                                startTime, endTime, otherMetaData, assemblyAPI)
-    writeJsonToFile(assemblyModel, new File(outFilePrefix + ".uaz.links.json"))
+    if (assemblyModel.isDefined)
+      writeJsonToFile(assemblyModel.get, new File(outFilePrefix + ".uaz.links.json"))
   }
+
 
   //
   // Private Methods
   //
 
+  /** Make and return the Sentence, Entity, Event, and (optionally) Assembly data models. */
+  def makeModels (paperId:String,
+                  allMentions:Seq[Mention],
+                  paperPassages:Seq[FriesEntry],
+                  startTime:Date,
+                  endTime:Date,
+                  assemblyApi: Option[Assembler] = None): (PropMap, PropMap, PropMap, Option[PropMap]) = {
+
+    // flatten mentions (per MITRE requirements), deduplicate, etc.
+    val sanitizedMentions = OutputDegrader.prepareForOutput(allMentions)
+
+    // dereference all coreference mentions:
+    val derefedMentions = sanitizedMentions.map(m => m.antecedentOrElse(m))
+
+    val otherMetaData = extractOtherMetaData(paperPassages)
+    val passageMap = passagesToMap(paperPassages) // map of FriesEntry, chunkId as key
+
+    // make Sentence Model
+    val sentModel = sentencesToModel(paperId, derefedMentions, passageMap,
+                                     startTime, endTime, otherMetaData)
+
+    // make Entity Map and Event Map: mappings from entities & events to unique ids
+    val (entityMap, eventMap) = makeEMaps(paperId, derefedMentions, passageMap)
+
+    // make Entity Model
+    val contextIdMap = new CtxIDed
+    val entityModel = entitiesToModel(paperId, derefedMentions, passageMap, entityMap,
+                                      contextIdMap, startTime, endTime, otherMetaData)
+
+    // make Event Model
+    val eventModel = eventsToModel(paperId, derefedMentions, passageMap, entityMap,
+                                   eventMap, contextIdMap, startTime, endTime,
+                                   otherMetaData, assemblyApi)
+
+    // make optional Assembly Model
+    val assemblyModel =
+      if (assemblyApi.isDefined)
+        Some(assemblyToModel(paperId, derefedMentions, passageMap, entityMap,
+                             eventMap, startTime, endTime, otherMetaData, assemblyApi.get))
+      else None
+
+    (sentModel, entityModel, eventModel, assemblyModel)
+  }
+
+
   /** Add fixed pieces of meta data to the given model, along with additional meta data
       from the given Map. */
-  private def addMetaInfo(model:PropMap, paperId:String, startTime:Date, endTime:Date,
-                          otherMetaData:Map[String, String]): Unit = {
+  private def addMetaInfo (model: PropMap,
+                           paperId: String,
+                           startTime: Date,
+                           endTime: Date,
+                           otherMetaData: Map[String, String]): Unit = {
     model("object-type") = "frame-collection"
 
     val meta = new PropMap
@@ -198,115 +197,119 @@ class FriesOutput extends JsonOutputter {
     model("object-meta") = meta
   }
 
-  /** Returns a 2-tuple of a model object, representing all entity mentions extracted
-      from this paper, and a map from entity pointers to unique IDs. */
-  private def entitiesToModel (paperId:String,
-                               allMentions:Seq[Mention],
-                               passageMap:Map[String, FriesEntry],
-                               contextIdMap:CtxIDed,
-                               startTime:Date,
-                               endTime:Date,
-                               otherMetaData:Map[String, String]): (PropMap, IDed) = {
+
+  /** Return an Assembly Model, representing selected assembly information for this paper. */
+  private def assemblyToModel (paperId: String,
+                               mentions: Seq[Mention],
+                               passageMap: Map[String, FriesEntry],
+                               entityMap: IDed,
+                               eventMap: IDed,
+                               startTime: Date,
+                               endTime: Date,
+                               otherMetaData: Map[String, String],
+                               assemblyApi: Assembler): PropMap = {
     val model:PropMap = new PropMap
     addMetaInfo(model, paperId, startTime, endTime, otherMetaData)
-    val entityMap = new IDed
 
-    // this stores all frames in this output
     val frames = new FrameList
     model("frames") = frames
 
-    // dereference all coreference mentions:
-    val derefedMentions = allMentions.map(m => m.antecedentOrElse(m))
+    for (mention <- eventMap.keys.toList.sortBy(m => (m.sentence, m.tokenInterval.start))) {
+      val fromId = lookupMentionId(mention, entityMap, eventMap)
+      if (fromId.isDefined) {
+        frames ++= make1toMLinkFrames(paperId, mention, fromId.get,
+                                      assemblyApi.getEquivalenceLinks(mention),
+                                      "equivalent-to", passageMap, entityMap, eventMap)
+        frames ++= make1to1LinkFrames(paperId, mention, fromId.get,
+                                      assemblyApi.getCausalPredecessors(mention),
+                                      "preceded-by", passageMap, entityMap, eventMap)
+      }
+    }
+    model                              // return the constructed assembly model
+  }
 
-    for(mention <- derefedMentions) {
+
+  /** Return an Entity Model, representing all entity mentions extracted from this paper. */
+  private def entitiesToModel (paperId: String,
+                               mentions: Seq[Mention],
+                               passageMap: Map[String, FriesEntry],
+                               entityMap: IDed,
+                               contextIdMap: CtxIDed,
+                               startTime: Date,
+                               endTime: Date,
+                               otherMetaData: Map[String, String]): PropMap = {
+    val model:PropMap = new PropMap
+    addMetaInfo(model, paperId, startTime, endTime, otherMetaData)
+
+    val frames = new FrameList
+    model("frames") = frames
+
+    for (mention <- mentions) {
       mention match {
         case em:BioTextBoundMention =>
           val passage = getPassageForMention(passageMap, em)
-          frames ++= mkEntityMention(paperId, passage, em, contextIdMap, entityMap)
-        case _ => // these are events; we will export them later
+          frames ++= makeEntityMention(paperId, passage, em, entityMap, contextIdMap)
+        case _ =>                           // these are events or relations
       }
     }
-    (model, entityMap)
+    model
   }
 
 
   /** Returns a model object representing all event mentions extracted from this paper. */
-  private def eventsToModel (
-    paperId:String,
-    allMentions:Seq[Mention],
-    passageMap:Map[String, FriesEntry],
-    contextIdMap:CtxIDed,
-    entityMap:IDed,
-    startTime:Date,
-    endTime:Date,
-    otherMetaData:Map[String, String],
-    assemblyAPI: Option[Assembler] = None): (PropMap, IDed) =
-  {
+  private def eventsToModel (paperId: String,
+                             mentions: Seq[Mention],
+                             passageMap: Map[String, FriesEntry],
+                             entityMap: IDed,
+                             eventMap: IDed,
+                             contextIdMap: CtxIDed,
+                             startTime: Date,
+                             endTime: Date,
+                             otherMetaData: Map[String, String],
+                             assemblyApi: Option[Assembler] = None): PropMap = {
     val model:PropMap = new PropMap
     addMetaInfo(model, paperId, startTime, endTime, otherMetaData)
 
-    // this stores all frames in this output
     val frames = new FrameList
     model("frames") = frames
 
-    // stores all ids for simple events
-    val eventMap = new IDed
-
-    // dereference all coreference mentions:
-    val derefedMentions = allMentions.map(m => m.antecedentOrElse(m))
-
-    // keeps just events:
-    val eventMentions = derefedMentions.filter(MentionManager.isEventMention)
-
-    // first, print all non regulation events
+    val eventMentions = mentions.filter(isEventMention) // only process events
     for (mention <- eventMentions) {
-      if (!mention.matches("Regulation")) {
-        val passage = getPassageForMention(passageMap, mention)
-        frames ++= mkEventMention(paperId, passage, mention.toBioMention, contextIdMap, entityMap, eventMap, assemblyAPI)
-      }
+      val passage = getPassageForMention(passageMap, mention)
+      frames ++= makeEventMention(paperId, passage, mention.toBioMention, entityMap,
+                                  eventMap, contextIdMap, assemblyApi)
     }
-
-    // now, print all regulation events, which control the above events
-    for (mention <- eventMentions) {
-      if (mention.matches("Regulation")) {
-        // "reach down" to process any child regulations first, before processing current regulation
-        val controlledRegulations = mention.arguments.getOrElse("controlled", Nil).filter(_.matches("Regulation"))
-        val controllerRegulations = mention.arguments.getOrElse("controller", Nil).filter(_.matches("Regulation"))
-        val regulations = controlledRegulations ++ controllerRegulations
-        // FIXME this is a hack to process the arguments of nested regulations
-        val childrenRegulations = regulations.flatMap(_.arguments.values.flatten).filter(_.matches("Regulation"))
-        for (reg <- childrenRegulations ++ regulations) {
-          val passage = getPassageForMention(passageMap, reg)
-          frames ++= mkEventMention(paperId, passage, reg.toBioMention, contextIdMap, entityMap, eventMap, assemblyAPI)
-        }
-        // process current regulation
-        val passage = getPassageForMention(passageMap, mention)
-        frames ++= mkEventMention(paperId, passage, mention.toBioMention, contextIdMap, entityMap, eventMap, assemblyAPI)
-      }
-    }
-    (model, eventMap)
+    model
   }
 
-  private def getChunkId(m:Mention):String = {
-    assert(m.document.id.isDefined, s"object ${m} has no document ID")
 
-    val did = m.document.id.get
+  /** Return the chunk ID for the given mention. */
+  private def getChunkId (mention:Mention): String = {
+    assert(mention.document.id.isDefined, s"object ${mention} has no document ID")
     // the chunk id is the string following the underscore in the document ids
+    val did = mention.document.id.get
     val chunkId = did.substring(did.lastIndexOf("_") + 1)
     chunkId
   }
 
   /** Return the passage selected by the given mention's chunk ID. */
-  private def getPassageForMention (passageMap:Map[String, FriesEntry],
-                                    mention:Mention): FriesEntry = {
+  private def getPassageForMention (passageMap: Map[String, FriesEntry],
+                                    mention: Mention): FriesEntry = {
     val chunkId = getChunkId(mention)
     assert(passageMap.contains(chunkId), s"passageMap missing chunkId ${chunkId}")
     passageMap.get(chunkId).get
   }
 
+  /** Return the unique id string for the given mention, fetched from the given map. */
+  private def getUniqueId (eMap:IDed, m: Mention): String = {
+    val uid = eMap.get(m)
+    assert(uid.isDefined, s"Should never happen: Mention ${m} has no unique ID")
+    uid.get
+  }
 
-  /** Lookup the context for the given mention. Return a newly created context ID and frame;
-      or an existing context ID,=; or neither, if the mention has no context. */
+
+  /** Lookup the context for the given mention. Return a newly created context ID and frame,
+      or an existing context ID, or neither if the mention has no context. */
   private def lookupContext (paperId: String,
                              passage: FriesEntry,
                              mention: BioMention,
@@ -321,7 +324,7 @@ class FriesOutput extends JsonOutputter {
       if (contextId.isEmpty) {           // if this is a new context
         val ctxid = mkContextId(paperId, passage, mention.sentence) // generate new context ID
         contextIdMap.put(context, ctxid)      // save the new context ID keyed by the context
-        contextFrame = Some(mkContextFrame(paperId, passage, mention, ctxid, context))
+        contextFrame = Some(makeContextFrame(paperId, passage, mention, ctxid, context))
         contextId = Some(ctxid)               // save the new context ID
       }
     }
@@ -329,197 +332,201 @@ class FriesOutput extends JsonOutputter {
     (contextId, contextFrame)
   }
 
+
+  /** Lookup the unique ID, for the given mention, in the given maps, based on the type
+      of the mention. Returns an option for the ID found, or None if no ID found. */
   private def lookupMentionId (mention: Mention,
-                               entityMap:IDed,
-                               eventMap:IDed): Option[String] = mention match {
+                               entityMap: IDed,
+                               eventMap: IDed): Option[String] = mention match {
     case tbm: TextBoundMention => entityMap.get(tbm)
     case evm: EventMention => eventMap.get(evm)
+    case rem: RelationMention => eventMap.get(rem)
     case _ => None
   }
 
-  private def mkContextId(paperId:String, passage:FriesEntry, offset:Int):String =
-    s"cntx-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${contextIdCntr.genNextId()}"
 
-  private def mkEntityId(paperId:String, passage:FriesEntry, offset:Int):String =
-    s"ment-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${entityIdCntr.genNextId()}"
-
-  private def mkEventId(paperId:String, passage:FriesEntry, offset:Int):String =
-    s"evem-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${eventIdCntr.genNextId()}"
-
-  private def mkLinkId(paperId:String, passage:FriesEntry, offset:Int):String =
-    s"link-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${linkIdCntr.genNextId()}"
-
-  private def mkPassageId(paperId:String, passage:FriesEntry):String =
-    s"pass-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}"
-
-  private def mkSentenceId(paperId:String, passage:FriesEntry, offset:Int):String =
-    s"sent-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset"
-
-
-  private def mkArgument(
-    name:String,
-    arg:Mention,
-    argIndex:Int,
-    entityMap: IDed,
-    eventMap:IDed,
-    argFeatures:Option[RoleWithFeatures] = None
-  ): PropMap =
-  {
-    val m = new PropMap
-    m("object-type") = "argument"
-    m("type") = prettifyLabel(name)
+  private def makeArgument (name: String,
+                            arg: Mention,
+                            argIndex: Int,
+                            parent: Mention,
+                            entityMap: IDed,
+                            eventMap: IDed,
+                            argFeatures:Option[RoleWithFeatures] = None): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "argument"
+    pm("type") = prettifyLabel(name)
     val argType = mkArgType(arg)
-    m("argument-type") = argType
-    m("index") = argIndex
-    m("text") = arg.text
+    pm("argument-type") = argType
+    pm("index") = argIndex
+    pm("text") = arg.text
 
     argType match {
-      case "complex" =>
-        // this is a complex: print the participants
-        assert(arg.isInstanceOf[RelationMention], "complex 'arg' is not an instance of RelationMention")
+
+      case "complex" =>                     // a Protein Complex: print the participants
+        if (!isRelationMention(arg))
+          throw new RuntimeException(s"Complex argument [${arg.text}] is not an instance of RelationMention")
         val participants = new PropMap
-        val complexParticipants = arg.asInstanceOf[RelationMention].arguments
-        for(key <- complexParticipants.keySet) {
-          // FIXME: resolve each participant.  Should this be done elsewhere?
-          val ms: Seq[Mention] = complexParticipants.get(key).get.map(m => m.antecedentOrElse(m))
-          for ((p, i) <- ms.zipWithIndex) {
-            assert(p.isInstanceOf[TextBoundMention], s"complex participant is not an instance of TextBoundMention: ${p}")
-            if (!entityMap.contains(p)) {
-              throw new RuntimeException(s"Complex participant [${p.text} [mods: ${p.toCorefMention.modifications.map(_.toString).mkString(" ")}}]] not in entityMap \nin event [$currEvent] \nin sentence[${p.document.sentences(p.sentence).words.mkString(" ")}]:\n" + p.json(pretty = true))
-            }
+        val complexArgs = arg.asInstanceOf[RelationMention].arguments
+        for (key <- complexArgs.keys.toList.sorted) {
+          val argSeq = complexArgs.get(key).get
+          for ((p, i) <- argSeq.zipWithIndex) {
+            if (!isTextBoundMention(p))
+              throw new RuntimeException(s"Complex participant [${p}] is not an instance of TextBoundMention")
+            if (!entityMap.contains(p))
+              throw new RuntimeException(s"Complex participant [${p.text} [mods: ${p.toCorefMention.modifications.map(_.toString).mkString(" ")}}]] not in entityMap \nin event [$parent.text] \nin sentence[${p.document.sentences(p.sentence).words.mkString(" ")}]:\n" + p.json(pretty = true))
+            // participant checks have passed
             participants(s"$key${i + 1}") = entityMap.get(p).get
           }
         }
-        m("args") = participants
+        pm("args") = participants
 
-      case "entity" =>
-        // this is an entity: fetch its id from the entity map
-        if(! entityMap.contains(arg)) {
-          /*
-          println("Document:")
-          for(i <- arg.document.sentences.indices) {
-            println(s"Sentence #$i: ${arg.document.sentences(i).words.mkString(" ")}")
-          }
-          println("entityMap:")
-          for(e <- entityMap.keySet) {
-            println(s"\t${e.text} with labels ${e.labels.mkString(", ")}")
-          }
-          */
-          throw new RuntimeException(s"Entity argument [${arg.text} [mods: ${arg.toCorefMention.modifications.map(_.toString).mkString(" ")}}]] not in entityMap \nin event [$currEvent] \nin sentence[${arg.document.sentences(arg.sentence).words.mkString(" ")}]:\n" + arg.json(pretty = true))
+      case "entity" =>                      // an Entity: fetch its ID from the entity map
+        if (!entityMap.contains(arg)) {
+          throw new RuntimeException(s"Entity argument [${arg.text} [mods: ${arg.toCorefMention.modifications.map(_.toString).mkString(" ")}]] not in entityMap \nin event [$parent.text] \nin sentence[${arg.document.sentences(arg.sentence).words.mkString(" ")}]:\n" + arg.json(pretty = true))
         }
-        m("arg") = entityMap.get(arg).get
+        pm("arg") = entityMap.get(arg).get
 
         // output any participant features associated with this entity by assembly:
         if (argFeatures.isDefined) {
-          val features = mkParticipantFeatures(argFeatures.get)
+          val features = makeParticipantFeatures(argFeatures.get)
           if (features.nonEmpty)
-            m("participant-features") = features
+            pm("participant-features") = features
         }
 
-      case "event" =>
-        // this is an event, which we MUST have seen before
-        // if (! eventMap.contains(arg)) {
-        //   println("CANNOT FIND ARG: " + arg + " with HASH CODE: " + arg.hashCode())
-        //   displayMention(arg)
-        // }
+      case "event" =>                       // an Event: fetch its ID from the event map
         if (!eventMap.contains(arg)) {
-          val msg = s"$arg with labels (${arg.labels}) found by ${arg.foundBy} is missing from the eventMap"
-          throw new Exception(msg)
+          throw new RuntimeException(s"Event argument [${arg}] with labels (${arg.labels}) found by ${arg.foundBy} is missing from the eventMap")
         }
-        m("arg") = eventMap.get(arg).get
+        pm("arg") = eventMap.get(arg).get
 
+      // default: should never happen!
       case _ =>
-        // we should never be here!
-        throw new RuntimeException("ERROR: unknown event type: " + arg.labels)
+        throw new RuntimeException("ERROR: unknown argument type in makeArgument: " + arg.labels)
     }
-    m
-  }
-
-  private def mkAssemblyModel (paperId:String,
-                               allMentions:Seq[Mention],
-                               passageMap:Map[String, FriesEntry],
-                               entityMap:IDed,
-                               eventMap:IDed,
-                               startTime:Date,
-                               endTime:Date,
-                               otherMetaData:Map[String, String],
-                               assemblyAPI: Assembler): PropMap = {
-    val assemblyModel:PropMap = new PropMap
-    addMetaInfo(assemblyModel, paperId, startTime, endTime, otherMetaData)
-    val frames = new FrameList
-    assemblyModel("frames") = frames
-
-    for (mention <- eventMap.keys.toList.sortBy(m => (m.sentence, m.tokenInterval.start))) {
-      val fromId = lookupMentionId(mention, entityMap, eventMap)
-      if (fromId.isDefined) {
-        frames ++= mk1toMLinkFrames(paperId, mention, fromId.get,
-                                    assemblyAPI.getEquivalenceLinks(mention), "equivalent-to",
-                                    passageMap, entityMap, eventMap)
-        frames ++= mk1to1LinkFrames(paperId, mention, fromId.get,
-                                    assemblyAPI.getCausalPredecessors(mention), "preceded-by",
-                                    passageMap, entityMap, eventMap)
-      }
-    }
-
-    assemblyModel                           // return the constructed assembly model
+    pm
   }
 
 
-  private def mkContextFrame (paperId:String,
-                              passage:FriesEntry,
-                              mention:BioMention,
-                              contextId:String,
-                              context:ContextMap): PropMap = {
+  /** Create and return a new context frame using the given mention and context info. */
+  private def makeContextFrame (paperId:String,
+                                passage:FriesEntry,
+                                mention:BioMention,
+                                contextId:String,
+                                context:ContextMap): PropMap = {
     val f = startFrame()
     f("frame-type") = "context"
     f("frame-id") = contextId
     f("scope") = mkSentenceId(paperId, passage, mention.sentence)
-    f("facets") = mkContextFacets(context)
+    f("facets") = makeContextFacets(context)
     f
   }
 
-  private def mkContextFacets (context: ContextMap): PropMap = {
-    val m = new PropMap
-    m("object-type") = "facet-set"
+  /** Create and return a new map containing the context facets from the given context info. */
+  private def makeContextFacets (context: ContextMap): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "facet-set"
     context.foreach { case(k, v) =>
       val friesFacetName = contextNameMap.getOrElse(k, k)
-      m(friesFacetName) = v
+      pm(friesFacetName) = v
     }
-    m
+    pm
   }
 
-  private def mkEntityMention(paperId:String,
-                              passageMeta:FriesEntry,
-                              mention:BioTextBoundMention,
-                              contextIdMap:CtxIDed,
-                              entityMap: IDed): List[PropMap] = {
+
+  /** Return a tuple of maps mapping entities to unique IDs and events to unique IDs. */
+  private def makeEMaps (paperId: String,
+                         mentions: Seq[Mention],
+                         passageMap: Map[String, FriesEntry]): (IDed, IDed) = {
+    val entityMap = new IDed
+    val eventMap = new IDed
+
+    for (mention <- mentions) {
+      mention match {
+        case em:BioTextBoundMention =>
+          makeEntityMapEntry(paperId, em, passageMap, entityMap)
+
+        case evm:BioEventMention =>
+          makeEventMapEntry(paperId, evm, passageMap, entityMap, eventMap)
+
+        case rem:BioRelationMention =>
+          makeEventMapEntry(paperId, rem, passageMap, entityMap, eventMap)
+
+        case _ => println(s"(makeEMaps): UNTYPED=${mention.toString()}") // DEBUGGING
+      }
+    }
+
+    (entityMap, eventMap)
+  }
+
+  /** Add an entry to the given Entity Map for the given entity mention. */
+  private def makeEntityMapEntry (paperId: String,
+                                  mention: BioTextBoundMention,
+                                  passageMap: Map[String, FriesEntry],
+                                  entityMap: IDed): Unit = {
+    val passage = getPassageForMention(passageMap, mention)
+    entityMap += mention -> mkEntityId(paperId, passage, mention.sentence)
+  }
+
+  /** Add an entry to the given Event Map for the given event mention. Then,
+      recursively add the event's arguments to the appropriate maps. */
+  private def makeEventMapEntry (paperId: String,
+                                 mention: BioMention,
+                                 passageMap: Map[String, FriesEntry],
+                                 entityMap: IDed,
+                                 eventMap: IDed): Unit = {
+    val passage = getPassageForMention(passageMap, mention)
+    eventMap += mention -> mkEventId(paperId, passage, mention.sentence)
+    mention.arguments.foreach {
+      case (_, argMentions) => argMentions.foreach { argMention =>
+        argMention match {
+          case evm:BioEventMention =>
+            makeEventMapEntry(paperId, evm, passageMap, entityMap, eventMap)
+
+          case rem:BioRelationMention =>
+            makeEventMapEntry(paperId, rem, passageMap, entityMap, eventMap)
+
+          case em:BioTextBoundMention =>
+            makeEntityMapEntry(paperId, em, passageMap, entityMap)
+
+          case _ => println(s"Argument Mention UNTYPED=${argMention.toString()}") // REMOVE LATER
+        }
+      }
+    }
+  }
+
+
+  /** Create and return a new entity mention frame for the given entity mention
+      and, possibly, a related context frame. */
+  private def makeEntityMention (paperId:String,
+                                 passageMeta:FriesEntry,
+                                 mention:BioTextBoundMention,
+                                 entityMap: IDed,
+                                 contextIdMap:CtxIDed): List[PropMap] = {
     val f = startFrame()
     f("frame-type") = "entity-mention"
-    val eid = mkEntityId(paperId, passageMeta, mention.sentence)
-    entityMap += mention -> eid
-    f("frame-id") = eid
+    f("frame-id") = getUniqueId(entityMap, mention)
     f("sentence") = mkSentenceId(paperId, passageMeta, mention.sentence)
-    f("start-pos") = mkRelativePosition(paperId, passageMeta, mention.startOffset)
-    f("end-pos") = mkRelativePosition(paperId, passageMeta, mention.endOffset)
+    f("start-pos") = makeRelativePosition(paperId, passageMeta, mention.startOffset)
+    f("end-pos") = makeRelativePosition(paperId, passageMeta, mention.endOffset)
     f("text") = mention.text
     f("type") = prettifyLabel(mention.displayLabel)
     val groundings = new FrameList
-    mention.grounding.foreach(grnd => groundings += mkGrounding(grnd))
+    mention.grounding.foreach(grnd => groundings += makeGrounding(grnd))
     f("xrefs") = groundings
     if (mention.isModified) {
       val ms = new FrameList
-      for(m <- mention.modifications) {
+      for (m <- mention.modifications) {
         m match {
-          case ptm: PTM => ms += mkPTM(ptm)
-          case m: Mutant => ms += mkMutant(m)
-          case _ => // we do not export anything else
+          case ptm: PTM => ms += makePTM(ptm)
+          case m: Mutant => ms += makeMutant(m)
+          case _ =>                         // we do not export anything else
         }
       }
       if (ms.nonEmpty)
         f("modifications") = ms
     }
 
-    // TODO (optional): add "index", i.e., the sentence-local number for this mention from this component
+    // OPTIONAL TODO: add "index": the sentence-local number for this mention from this component.
 
     // handle context features:
     val (contextId, contextFrame) = lookupContext(paperId, passageMeta, mention, contextIdMap)
@@ -535,30 +542,22 @@ class FriesOutput extends JsonOutputter {
       List(f)
   }
 
-  private def mkEventMention(
-    paperId:String,
-    passageMeta:FriesEntry,
-    mention:BioMention,
-    contextIdMap: CtxIDed,
-    entityMap: IDed,
-    eventMap:IDed,
-    assemblyAPI: Option[Assembler] = None): List[PropMap] =
-  {
-    currEvent = mention.text
 
+  /** Create and return a new event mention frame for the given event mention
+      and, possibly, a related context frame. */
+  private def makeEventMention (paperId: String,
+                                passageMeta: FriesEntry,
+                                mention: BioMention,
+                                entityMap: IDed,
+                                eventMap: IDed,
+                                contextIdMap: CtxIDed,
+                                assemblyApi: Option[Assembler] = None): List[PropMap] = {
     val f = startFrame()
     f("frame-type") = "event-mention"
-    val eid = mkEventId(paperId, passageMeta, mention.sentence)
-    eventMap += mention -> eid
-
-    //println(s"ADDED EVENT: $mention -> $eid")
-    //println(s"HASH CODE: " + mention.hashCode())
-    //displayMention(mention)
-
-    f("frame-id") = eid
+    f("frame-id") = getUniqueId(eventMap, mention)
     f("sentence") = mkSentenceId(paperId, passageMeta, mention.sentence)
-    f("start-pos") = mkRelativePosition(paperId, passageMeta, mention.startOffset)
-    f("end-pos") = mkRelativePosition(paperId, passageMeta, mention.endOffset)
+    f("start-pos") = makeRelativePosition(paperId, passageMeta, mention.startOffset)
+    f("end-pos") = makeRelativePosition(paperId, passageMeta, mention.endOffset)
     f("text") = mention.text
     if (mention.isInstanceOf[BioEventMention]) {
       f("trigger") = mention.asInstanceOf[BioEventMention].trigger.text
@@ -569,29 +568,25 @@ class FriesOutput extends JsonOutputter {
 
     val evType = mkEventType(mention)
     f("type") = evType
-    if(evType != "complex-assembly")
+    if (evType != "complex-assembly")
       f("subtype") = prettifyLabel(mention.label)
 
-    var arguments:Option[Map[String, Seq[Mention]]] = None
-    mention match {
-      case em:BioEventMention => arguments = Some(em.arguments)
-      case rm:BioRelationMention => arguments = Some(rm.arguments)
-    }
-
     // if using the Assembly subsystem, get all input participant features for this event
-    val inputFeatures = assemblyAPI.map(_.getInputFeaturesByParticipants(mention))
+    val inputFeatures = assemblyApi.map(_.getInputFeaturesByParticipants(mention))
 
-    // handle event arguments
+    // process the arguments if present:
+    var arguments: Option[Map[String, Seq[Mention]]] =
+      if (isEventMention(mention)) Some(mention.arguments) else None
+
     if (arguments.isDefined) {
       val argList = new FrameList
 
-      // process the arguments:
-      for (key <- arguments.get.keys) {
+      for (key <- arguments.get.keys.toList.sorted) {
         arguments.get.get(key).foreach { argSeq =>
           for (i <- argSeq.indices) {
-            val derefedArg = argSeq(i).antecedentOrElse(argSeq(i).toCorefMention)
-            val argFeatures = inputFeatures.flatMap(_.get(derefedArg))
-            argList += mkArgument(key, derefedArg, i, entityMap, eventMap, argFeatures)
+            val ithArg = argSeq(i)
+            val argFeatures = inputFeatures.flatMap(_.get(ithArg))
+            argList += makeArgument(key, ithArg, i, mention, entityMap, eventMap, argFeatures)
           }
         }
       }
@@ -599,13 +594,12 @@ class FriesOutput extends JsonOutputter {
     }
 
     // event modifications
-    if (MentionManager.isNegated(mention))
+    if (isNegated(mention))
       f("is-negated") = true
-    if (MentionManager.isHypothesized(mention))
+    if (isHypothesized(mention))
       f("is-hypothesis") = true
 
-    // TODO (optional): add "index", i.e., the sentence-local number for this mention
-    //                  from this component.
+    // OPTIONAL TODO: add "index": the sentence-local number for this mention from this component.
 
     // handle context features:
     val (contextId, contextFrame) = lookupContext(paperId, passageMeta, mention, contextIdMap)
@@ -622,32 +616,33 @@ class FriesOutput extends JsonOutputter {
   }
 
 
-  private def mkGrounding (grounding:KBResolution): PropMap = {
-    val m = new PropMap
-    m("object-type") = "db-reference"
-    m("namespace") = grounding.namespace
-    m("id") = grounding.id
-    m
+  /** Create and return a single grounding map. */
+  private def makeGrounding (grounding:KBResolution): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "db-reference"
+    pm("namespace") = grounding.namespace
+    pm("id") = grounding.id
+    pm
   }
 
-
-  private def mkLinkArgumentFrame (arg:String,
-                                   argType:String,
-                                   index:Option[Integer]=None): PropMap = {
-    val m = new PropMap
-    m("object-type") = "argument"
-    m("type") = argType
-    m("arg") = arg
+  /** Create and return a single assembly link argument map. */
+  private def makeLinkArgumentFrame (arg: String,
+                                     argType: String,
+                                     index: Option[Integer] = None): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "argument"
+    pm("type") = argType
+    pm("arg") = arg
     if (index.isDefined)
-      m("index") = index.get
-    m
+      pm("index") = index.get
+    pm
   }
 
   /** Create and return a new binary Link frame. */
-  private def mkLinkFrame (linkId: String,
-                           frameType: String,
-                           foundBy: String,
-                           args: List[PropMap]): PropMap = {
+  private def makeLinkFrame (linkId: String,
+                             frameType: String,
+                             foundBy: String,
+                             args: List[PropMap]): PropMap = {
     val f = startFrame()
     f("frame-id") = linkId
     f("frame-type") = "link"
@@ -661,7 +656,7 @@ class FriesOutput extends JsonOutputter {
   }
 
   /** Return a list of link frames made from the given mention and set of assembly links. */
-  private def mk1to1LinkFrames [L <: AssemblyLink] (
+  private def make1to1LinkFrames [L <: AssemblyLink] (
     paperId: String,
     from: Mention,
     fromId: String,
@@ -680,9 +675,9 @@ class FriesOutput extends JsonOutputter {
           // validate: from == after, "link should represent predecessors of 'from' Mention"
           val m2Id = lookupMentionId(before, entityMap, eventMap)
           if (m2Id.isDefined) {
-            val args = List(mkLinkArgumentFrame(fromId, "after"),
-                            mkLinkArgumentFrame(m2Id.get, "before"))
-            frames += mkLinkFrame(linkId, frameType, foundBy, args)
+            val args = List(makeLinkArgumentFrame(fromId, "after"),
+                            makeLinkArgumentFrame(m2Id.get, "before"))
+            frames += makeLinkFrame(linkId, frameType, foundBy, args)
           }
         case unknown =>
           System.err.println(s"Cannot create 1-to-1 links for type '${unknown.getClass}'")
@@ -692,7 +687,7 @@ class FriesOutput extends JsonOutputter {
   }
 
   /** Return a singleton list of link frame made from the given mention and set of assembly links. */
-  private def mk1toMLinkFrames [L <: AssemblyLink] (
+  private def make1toMLinkFrames [L <: AssemblyLink] (
     paperId: String,
     from: Mention,
     fromId: String,
@@ -711,13 +706,13 @@ class FriesOutput extends JsonOutputter {
       links.head match {
         case Equivalence(_, m2, _) =>
           val args: ListBuffer[PropMap] = new ListBuffer()
-          args += mkLinkArgumentFrame(fromId, "from") // add FROM as argument zero
+          args += makeLinkArgumentFrame(fromId, "from") // add FROM as argument zero
           args ++= links.zipWithIndex.flatMap { case(link:Equivalence, ndx:Int) =>
             lookupMentionId(link.m2, entityMap, eventMap).map(m2Id =>
-              mkLinkArgumentFrame(m2Id, s"to${ndx}"))
+              makeLinkArgumentFrame(m2Id, s"to${ndx}"))
           }
           if (args.size > 1)                  // must have at least one TO argument
-            frames += mkLinkFrame(linkId, frameType, foundBy, args.toList)
+            frames += makeLinkFrame(linkId, frameType, foundBy, args.toList)
 
         case unknown =>
           System.err.println(s"Cannot create 1-to-M links for type '${unknown.getClass}'")
@@ -727,37 +722,38 @@ class FriesOutput extends JsonOutputter {
   }
 
 
-  private def mkMutant(mutant: Mutant): PropMap = {
-    val m = new PropMap
-    m("object-type") = "modification"
-    m("type") = "mutation"
-    m("evidence") = mutant.evidence.text
-    m
+  /** Create and return a single mutant modification map. */
+  private def makeMutant (mutant: Mutant): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "modification"
+    pm("type") = "mutation"
+    pm("evidence") = mutant.evidence.text
+    pm
   }
 
 
   /** Return a list of participant features for the given role/argument features object. */
-  private def mkParticipantFeatures (rwfs: RoleWithFeatures): FrameList = {
-    (new FrameList) ++= rwfs.features.map{ f:AssemblyPTM => mkParticipantFeature(f) }
+  private def makeParticipantFeatures (rwfs: RoleWithFeatures): FrameList = {
+    (new FrameList) ++= rwfs.features.map{ f:AssemblyPTM => makeParticipantFeature(f) }
   }
 
   /** Return a participant feature map for the given modification object. */
-  private def mkParticipantFeature (ptm: AssemblyPTM): PropMap = {
-    val m = new PropMap
-    m("object-type") = "feature"
-    m("type") = ptm.label
-    m("negated") = ptm.negated
+  private def makeParticipantFeature (ptm: AssemblyPTM): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "feature"
+    pm("type") = ptm.label
+    pm("negated") = ptm.negated
     if (ptm.site.isDefined)
-      m("site") = ptm.site.get
-    m
+      pm("site") = ptm.site.get
+    pm
   }
 
 
-  /** Create and return a new Passage frame. */
-  private def mkPassage(model:PropMap,
-                        paperId:String,
-                        passage:FriesEntry,
-                        passageDoc:Document): PropMap = {
+  /** Create and return a new passage frame. */
+  private def makePassage (model: PropMap,
+                           paperId: String,
+                           passage: FriesEntry,
+                           passageDoc: Document): PropMap = {
     val f = startFrame(Some("nxml2fries"))
     f("frame-id") = mkPassageId(paperId, passage)
     f("frame-type") = "passage"
@@ -770,95 +766,120 @@ class FriesOutput extends JsonOutputter {
     f
   }
 
-  private def mkPTM (ptm: PTM): PropMap = {
-    val m = new PropMap
-    m("object-type") = "modification"
-    m("type") = ptm.label
-    m("negated") = ptm.negated
+  /** Create and return a single PTM modification map. */
+  private def makePTM (ptm: PTM): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "modification"
+    pm("type") = ptm.label
+    pm("negated") = ptm.negated
     if (ptm.site.isDefined) {
-      m("site") = ptm.site.get.text
+      pm("site") = ptm.site.get.text
     }
-    m
+    pm
   }
 
-  private def mkRelativePosition(paperId:String,
-                                 passageMeta:FriesEntry,
-                                 characterOffset:Int): PropMap = {
-    val pos = new PropMap
-    pos("object-type") = "relative-pos"
-    pos("reference") = mkPassageId(paperId, passageMeta)
-    pos("offset") = characterOffset
-    pos
+  /** Create and return a single relative position map. */
+  private def makeRelativePosition (paperId: String,
+                                  passageMeta: FriesEntry,
+                                  characterOffset: Int): PropMap = {
+    val pm = new PropMap
+    pm("object-type") = "relative-pos"
+    pm("reference") = mkPassageId(paperId, passageMeta)
+    pm("offset") = characterOffset
+    pm
   }
 
-  private def mkSentence(model:PropMap,
-                         paperId:String,
-                         passageMeta:FriesEntry,
-                         passageDoc:Document,
-                         offset:Int): PropMap = {
-    val sent = startFrame(Some("BioNLPProcessor"))
-    sent("frame-type") = "sentence"
-    sent("frame-id") = mkSentenceId(paperId, passageMeta, offset)
-    sent("passage") = mkPassageId(paperId, passageMeta)
-    sent("start-pos") = mkRelativePosition(paperId, passageMeta, MentionManager.sentenceStartCharacterOffset(passageDoc, offset))
-    sent("end-pos") = mkRelativePosition(paperId, passageMeta, MentionManager.sentenceEndCharacterOffset(passageDoc, offset))
-    sent("text") = passageDoc.sentences(offset).getSentenceText
-    sent
+  /** Create and return a sentence frame. */
+  private def makeSentence (model: PropMap,
+                            paperId: String,
+                            passageMeta: FriesEntry,
+                            passageDoc: Document,
+                            offset: Int): PropMap = {
+    val f = startFrame(Some("BioNLPProcessor"))
+    f("frame-type") = "sentence"
+    f("frame-id") = mkSentenceId(paperId, passageMeta, offset)
+    f("passage") = mkPassageId(paperId, passageMeta)
+    f("start-pos") = makeRelativePosition(paperId, passageMeta, sentenceStartCharacterOffset(passageDoc, offset))
+    f("end-pos") = makeRelativePosition(paperId, passageMeta, sentenceEndCharacterOffset(passageDoc, offset))
+    f("text") = passageDoc.sentences(offset).getSentenceText
+    f
   }
 
-  private def mkSentences(model:PropMap,
-                          paperId:String,
-                          passageMeta:FriesEntry,
-                          passageDoc:Document): Seq[PropMap] = {
+  /** Create and return a sequence of sentence frames. */
+  private def makeSentences (model: PropMap,
+                             paperId: String,
+                             passageMeta: FriesEntry,
+                             passageDoc: Document): Seq[PropMap] = {
     val sents = new ListBuffer[PropMap]
-    for(i <- passageDoc.sentences.indices) {
-      sents += mkSentence(model, paperId, passageMeta, passageDoc, i)
+    for (i <- passageDoc.sentences.indices) {
+      sents += makeSentence(model, paperId, passageMeta, passageDoc, i)
     }
     sents.toList
   }
 
 
+  private def mkContextId (paperId:String, passage:FriesEntry, offset:Int): String =
+    s"cntx-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${contextIdCntr.genNextId()}"
+
+  private def mkEntityId (paperId:String, passage:FriesEntry, offset:Int): String =
+    s"ment-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${entityIdCntr.genNextId()}"
+
+  private def mkEventId (paperId:String, passage:FriesEntry, offset:Int): String =
+    s"evem-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${eventIdCntr.genNextId()}"
+
+  private def mkLinkId (paperId:String, passage:FriesEntry, offset:Int): String =
+    s"link-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset-${linkIdCntr.genNextId()}"
+
+  private def mkPassageId (paperId:String, passage:FriesEntry): String =
+    s"pass-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}"
+
+  private def mkSentenceId (paperId:String, passage:FriesEntry, offset:Int): String =
+    s"sent-$paperId-$ORGANIZATION-$RUN_ID-${passage.chunkId}-$offset"
+
+
   /** Creates a map of all FriesEntries, using chunkId as key */
-  private def passagesToMap(paperPassages:Seq[FriesEntry]):Map[String, FriesEntry] = {
+  private def passagesToMap (paperPassages: Seq[FriesEntry]): Map[String, FriesEntry] = {
     val map = new mutable.HashMap[String, FriesEntry]()
-    for(e <- paperPassages) map += e.chunkId -> e
+    for (e <- paperPassages) map += e.chunkId -> e
     map.toMap
   }
 
+
   /** Returns a data-structure representing all sections and sentences in this paper. */
-  private def sentencesToModel (paperId:String,
-                                allMentions:Seq[Mention],
-                                passageMap:Map[String, FriesEntry],
-                                startTime:Date,
-                                endTime:Date,
-                                otherMetaData:Map[String, String]): PropMap = {
+  private def sentencesToModel (paperId: String,
+                                mentions: Seq[Mention],
+                                passageMap: Map[String, FriesEntry],
+                                startTime: Date,
+                                endTime: Date,
+                                otherMetaData: Map[String, String]): PropMap = {
     val model:PropMap = new PropMap
     addMetaInfo(model, paperId, startTime, endTime, otherMetaData)
 
     // keeps track of all documents created for each entry
     val passageDocs = new mutable.HashMap[String, Document]()
-    for(m <- allMentions)  {
+    for (m <- mentions)  {
       val chunkId = getChunkId(m)
-      if(! passageDocs.contains(chunkId)) {
+      if (!passageDocs.contains(chunkId)) {
         passageDocs += chunkId -> m.document
       }
     }
 
-    // this stores all frames in this output
+    // store all frames in this output
     val frames = new FrameList
     model("frames") = frames
 
     // now output all passages as individual frames
-    for(chunkId <- passageDocs.keySet) {
+    for (chunkId <- passageDocs.keySet) {
       assert(passageMap.contains(chunkId), s"passageMap missing chunkId ${chunkId}")
-      frames += mkPassage(model, paperId, passageMap.get(chunkId).get, passageDocs.get(chunkId).get)
-      frames ++= mkSentences(model, paperId, passageMap.get(chunkId).get, passageDocs.get(chunkId).get)
+      frames += makePassage(model, paperId, passageMap.get(chunkId).get, passageDocs.get(chunkId).get)
+      frames ++= makeSentences(model, paperId, passageMap.get(chunkId).get, passageDocs.get(chunkId).get)
     }
     model
   }
 
 
-  private def startFrame (component:Option[String]=None): PropMap = {
+  /** Make and return a new Frame optionally initialized with component type. */
+  private def startFrame (component: Option[String] = None): PropMap = {
     val f = new PropMap
     f("object-type") = "frame"
     if (component.isDefined) {              // no need to override default meta-data each time
@@ -871,7 +892,7 @@ class FriesOutput extends JsonOutputter {
   }
 
   /** Convert the entire output data structure to JSON and write it to the given file. */
-  private def writeJsonToFile (model:PropMap, outFile:File) = {
+  private def writeJsonToFile (model: PropMap, outFile: File) = {
     val out:PrintWriter = new PrintWriter(new BufferedWriter(new FileWriter(outFile)))
     Serialization.writePretty(model, out)
     out.println()                           // add final newline which serialization omits
@@ -880,4 +901,3 @@ class FriesOutput extends JsonOutputter {
   }
 
 }
-
